fix(ME-83): Fixing code and CodeRabbit
