+++ app/Models/User.php
@@ -8,7 +8,7 @@ use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 use Illuminate\Foundation\Auth\User as Authenticatable;
 use Illuminate\Notifications\Notifiable;
 use Laravel\Sanctum\HasApiTokens;
-use MedineTech\Companies\Infrastructure\Persistence\Eloquent\CompanyModel;
+use MedineTech\Backoffice\Companies\Infrastructure\Persistence\Eloquent\CompanyModel;
 
 
 class User extends Authenticatable
