+++ routes/api.php
@@ -9,3 +9,4 @@ Route::middleware(['auth:sanctum'])->get('/user', function (Request $request) {
 
 require __DIR__.'/auth.php';
 require __DIR__.'/backoffice/first-companies.php';
+require __DIR__.'/backoffice/companies.php';
