+++ src/Backoffice/Companies/Application/Create/CompanyCreator.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace MedineTech\Companies\Application\Create;
+namespace MedineTech\Backoffice\Companies\Application\Create;
 
-use MedineTech\Companies\Domain\Company;
-use MedineTech\Companies\Domain\CompanyRepository;
+use MedineTech\Backoffice\Companies\Domain\Company;
+use MedineTech\Backoffice\Companies\Domain\CompanyRepository;
 
 final class CompanyCreator
 {
