+++ src/Backoffice/Companies/Infrastructure/Persistence/Eloquent/CompanyModel.php
@@ -2,9 +2,8 @@
 
 declare(strict_types=1);
 
-namespace MedineTech\Companies\Infrastructure\Persistence\Eloquent;
+namespace MedineTech\Backoffice\Companies\Infrastructure\Persistence\Eloquent;
 
-use Illuminate\Database\Eloquent\Model;
 use Stancl\Tenancy\Database\Models\Tenant;
 
 final class CompanyModel extends Tenant
