+++ tests/Backoffice/Companies/Application/Find/CompanyFinderTest.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace Tests\Backoffice\Companies\Application\Find;
 
-use MedineTech\Companies\Application\Find\CompanyFinder;
-use MedineTech\Companies\Application\Find\CompanyFinderRequest;
-use MedineTech\Companies\Domain\CompanyRepository;
+use MedineTech\Backoffice\Companies\Application\Find\CompanyFinder;
+use MedineTech\Backoffice\Companies\Application\Find\CompanyFinderRequest;
+use MedineTech\Backoffice\Companies\Domain\CompanyRepository;
 use MedineTech\Shared\Domain\ValueObject\Uuid;
 use PHPUnit\Framework\Attributes\Test;
 use Tests\Backoffice\Companies\CompanyUnitTestCase;
